python/requirements.txt
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
executing==1.2.0
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywin32==305
pyzmq==25.0.2
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
style==1.1.0
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.4.6
urllib3==1.26.15
wcwidth==0.2.6
websockets==10.4
wsproto==1.2.0
